+++ client/src/views/index.ts
@@ -1,5 +1,6 @@
 export { DashboardView } from "./DashboardView";
 export { CalculatorView } from "./CalculatorView";
+export { ProductionView } from "./ProductionView";
 export { HistoryView } from "./HistoryView";
 export { WarehouseView } from "./WarehouseView";
 export { ReportsView } from "./ReportsView";